Refactor GUI classes
